requirements.txt
mediapipe
opencv-python
numpy
ffmpeg-python